:recycle: relocate user entity
